Modified scripts to add some checks . Still in barebones version , need to add a help part and all . Now can choose not to commit to the git repo by using a "-g" switch on both the new and edit shell scripts

+++ initp.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
-rm posts
+if [ -f posts ]
+then
+	rm posts
+fi
+
 touch posts
 
 ls -l *.markdown | awk '{print $9}' | cut -d "." -f1 | while read LINE
